Add float XP support and update chess XP formula
- Change XP storage from int to float for fractional XP support
- Update chess XP formula: 0.1 XP/move + outcome bonus (5 win, 2 draw, 0 loss)
- Add formatXP helper to TypeScript skill components for clean display
- Update type hints in chain_state.py and skills_manager.py

+++ core/block.py
@@ -771,7 +771,7 @@ def create_game_block(
     total_moves: int,
     pgn: str,
     duration_seconds: int,
-    xp_earned: int,
+    xp_earned: float,
     key_moments: Optional[List[Dict[str, Any]]] = None,
     chat_log: Optional[List[Dict[str, Any]]] = None
 ) -> Block:
